requirements.txt
bandit>=1.7.10
black>=24.10.0
gunicorn==23.0.0
ipython>=8.28.0
isort>=5.13.2
pillow==10.4.0
pip-check-updates>=0.27.0
pre_commit>=4.0.1
prospector>=1.12.1
psycopg2_binary==2.9.10
pytest>=8.3.3
pytest-django>=4.9.0
pyupgrade>=3.18.0
requests==2.32.3
wagtail==6.2.2
